Implement Save/Restore state. Switch metrics to ring buffer.
Now possible to save EWMAs and restore from the saved EWMA state to
ensure that the CB's learned data is not lost.

TODO: Handle warmup on restore. Currently Levee restores directly into
operational state. This should be changed to deferred restoration
after warmup is completed.

The raw metrics were stored in a buffer that was reset to 0 when
full. This is now replaced with a continously updating ring
buffer. EWMA updates happen once per wrap-around vs. the old behaviour
of once per buffer reset.

Added some long running benchmarks to test the CB's behaviour against
a "Prescient" breaker that knows the test pattern ahead of time. This
can give a good indication of the breaker's adaptability against an ideal.